+++ lib/components/Scribe/extension/extension-selectedText.ts
