+++ lib/components/Scribe/extension/index.ts
