+++ lib/components/Scribe/extension/slashCommand/index.ts
