+++ lib/components/Scribe/index.tsx
